Tidy: Refactor assessment results controllers
Two controllers that are identical. One for assessment
results for passported one for non-passported.

Can also potentially rename the child controllers, steps
and perhaps look at the views to remove repetition.
